requirements.txt
Flask==1.1.2
pycryptodome==3.9.8
zeep==3.4.0
